HttpAuth mock for dev & testing
Added a middleware that mocks Http Basic Auth, setting the
request.META['REMOTE_USER'] to whatever is entered in the basic auth
dialog.
For testing, you must add a HTTP_AUTHENTICATION parameter to all testing
client requests that you need an authenticated user.
This does not yet work for the "first" login of a user, as the
RegistryUserBackend will look for the request.META keys 'mail' and
'Shib-Identity-Provider' to modify the new user. You will have to do
this manually for now.