initial_data deprecated - copied initial_data to campus.json and manually loading it now
Prior to 1.9, Django automatically loaded whatever was in the file 'initial_data.json' as part of the syncdb command. This behavior has been deprecated. The warning suggested making a migration, to populate the database, but since we are already loading data manually (collection.json), it seemed a better upgrade path to rename the file and add a line to the README under Load - may need to add this to deployment scripts later also