Removed the longerusernames app.
This app has not been maintained since 2012. The Django username maxlength was 30 chars prior to Django 1.10, so this app extended the maxlength for usernames to 255. Django 1.10 introduces a feature extending the maxlength for usernames to 150. I evaluated all usernames on registry production and registry stage, and we have none exceeding 150, so I belief it's safe to remove this dependency.