Django 1.10 added a check for is_active in RemoteUserBackend. AllowAllUsersRemoteUserBackend doesn't have the check.
Registry uses this check to differentiate between a requested account and an active account. Since Shibboleth handles identity, we know the user is authenticated, but they may not have an account with the registry yet. Thus they are 'authenticated' but inactive. They should still pass the user_can_authenticate() test, though.